+++ src/components/UserSetting.tsx
@@ -1,8 +1,7 @@
 import { Stack, Heading, Text, Spacer } from '@chakra-ui/react'
 import React from 'react'
-import { NumberInputForm } from '../Components'
+import { ExposureValueBox, NumberInputForm } from './Common'
 import { useAppSelector, useAppDispatch } from '../hooks'
-import { calculateExposureValue } from '../lib'
 import {
   changeUserIso,
   changeUserShutterSpeed,
@@ -23,7 +22,10 @@ export function UserSetting() {
       borderColor={'gray.300'}
       w={'xs'}
     >
-      <Heading>Your setting</Heading>
+      <Heading>
+        Your <br />
+        setting
+      </Heading>
       <Spacer />
       <NumberInputForm
         label={'ISO'}
@@ -43,7 +45,7 @@ export function UserSetting() {
         handler={(_, value) => dispatch(changeUserFNumber(value))}
         step={0.1}
       />
-      <Text>Exposure value(100) : {calculateExposureValue(exposure)}</Text>
+      <ExposureValueBox exposure={exposure} />
     </Stack>
   )
 }
